Refactor into smaller tests
