+++ packages/twenty-front/src/modules/views/hooks/__tests__/useViewBar.test.tsx
@@ -1,33 +1,28 @@
 import { act } from 'react-dom/test-utils';
 import { MemoryRouter, useSearchParams } from 'react-router-dom';
-import { gql } from '@apollo/client';
 import { MockedProvider } from '@apollo/client/testing';
 import { renderHook, waitFor } from '@testing-library/react';
 import { RecoilRoot, useRecoilState, useRecoilValue } from 'recoil';
 import { v4 as uuidv4 } from 'uuid';
 
 import { ObjectMetadataItem } from '@/object-metadata/types/ObjectMetadataItem';
-import { FieldMetadata } from '@/object-record/field/types/FieldMetadata';
-import { Filter } from '@/object-record/object-filter-dropdown/types/Filter';
-import { FilterDefinition } from '@/object-record/object-filter-dropdown/types/FilterDefinition';
-import { Sort } from '@/object-record/object-sort-dropdown/types/Sort';
-import { SortDefinition } from '@/object-record/object-sort-dropdown/types/SortDefinition';
-import { ColumnDefinition } from '@/object-record/record-table/types/ColumnDefinition';
 import { generateDeleteOneRecordMutation } from '@/object-record/utils/generateDeleteOneRecordMutation';
-import { getScopedFamilyStateDeprecated } from '@/ui/utilities/recoil-scope/utils/getScopedFamilyStateDeprecated';
 import { getScopedStateDeprecated } from '@/ui/utilities/recoil-scope/utils/getScopedStateDeprecated';
+import {
+  filterDefinition,
+  viewFilter,
+} from '@/views/hooks/__tests__/useViewBar_ViewFilters.test';
+import {
+  sortDefinition,
+  viewSort,
+} from '@/views/hooks/__tests__/useViewBar_ViewSorts.test';
 import { useViewScopedStates } from '@/views/hooks/internal/useViewScopedStates';
 import { useViewBar } from '@/views/hooks/useViewBar';
 import { ViewScopeInternalContext } from '@/views/scopes/scope-internal-context/ViewScopeInternalContext';
-import { currentViewFieldsScopedFamilyState } from '@/views/states/currentViewFieldsScopedFamilyState';
 import { entityCountInCurrentViewScopedState } from '@/views/states/entityCountInCurrentViewScopedState';
 import { viewEditModeScopedState } from '@/views/states/viewEditModeScopedState';
 import { viewObjectMetadataIdScopeState } from '@/views/states/viewObjectMetadataIdScopeState';
 import { viewTypeScopedState } from '@/views/states/viewTypeScopedState';
-import { ViewField } from '@/views/types/ViewField';
-import { ViewFilter } from '@/views/types/ViewFilter';
-import { ViewFilterOperand } from '@/views/types/ViewFilterOperand';
-import { ViewSort } from '@/views/types/ViewSort';
 import { ViewType } from '@/views/types/ViewType';
 
 import 'cross-fetch/polyfill';
@@ -55,30 +50,8 @@ const mocks = [
       data: { deleteView: { id: '' } },
     })),
   },
-  {
-    request: {
-      query: gql`
-        mutation CreateOneViewField($input: ViewFieldCreateInput!) {
-          createViewField(data: $input) {
-            id
-          }
-        }
-      `,
-      variables: {
-        input: {
-          fieldMetadataId: 'id',
-          viewId: 'mocked-uuid',
-          isVisible: true,
-          size: 1,
-          position: 1,
-        },
-      },
-    },
-    result: jest.fn(() => ({
-      data: { createViewField: { id: '' } },
-    })),
-  },
 ];
+
 const Wrapper = ({ children }: { children: React.ReactNode }) => (
   <MemoryRouter
     initialEntries={['/one', '/two', { pathname: '/three' }]}
@@ -102,53 +75,6 @@ const renderHookConfig = {
 };
 
 const viewBarId = 'viewBarTestId';
-
-const fieldsToPersist = [
-  {
-    fieldMetadataId: 'id',
-    label: 'label',
-    iconName: 'icon',
-    type: 'UUID',
-    isVisible: true,
-    size: 1,
-    position: 1,
-    metadata: {
-      positon: 0,
-      size: 2,
-      objectMetadataNameSingular: 'view',
-      fieldName: 'view',
-    },
-  },
-] as unknown as ViewField[];
-
-const filterDefinition: FilterDefinition = {
-  fieldMetadataId: '113ea8f8-1908-4c9c-9984-3f23c96b92f5',
-  label: 'label',
-  iconName: 'iconName',
-  type: 'TEXT',
-};
-
-const viewFilter: ViewFilter = {
-  id: 'id',
-  fieldMetadataId: '113ea8f8-1908-4c9c-9984-3f23c96b92f5',
-  operand: ViewFilterOperand.Is,
-  value: 'value',
-  displayValue: 'displayValue',
-  definition: filterDefinition,
-};
-
-const sortDefinition: SortDefinition = {
-  fieldMetadataId: '12ecdf87-506f-44a7-98c6-393e5f05b225',
-  label: 'label',
-  iconName: 'icon',
-};
-
-const viewSort: ViewSort = {
-  id: '88930a16-685f-493b-a96b-91ca55666bba',
-  fieldMetadataId: '12ecdf87-506f-44a7-98c6-393e5f05b225',
-  direction: 'asc',
-  definition: sortDefinition,
-};
 
 describe('useViewBar', () => {
   it('should set and get current view Id', () => {
@@ -346,433 +272,6 @@ describe('useViewBar', () => {
     });
 
     expect(result.current.count).toBe(1);
-  });
-
-  it('should update current fields', async () => {
-    const { result } = renderHook(
-      () => ({
-        viewBar: useViewBar({ viewBarId }),
-        currentFields: useRecoilState(
-          getScopedFamilyStateDeprecated(
-            currentViewFieldsScopedFamilyState,
-            viewBarId,
-            mockedUuid,
-          ),
-        )[0],
-      }),
-      renderHookConfig,
-    );
-
-    expect(result.current.currentFields).toStrictEqual([]);
-    await act(async () => {
-      result.current.viewBar.setCurrentViewId(mockedUuid);
-      result.current.viewBar.setViewObjectMetadataId('newId');
-      result.current.viewBar.persistViewFields(fieldsToPersist);
-    });
-
-    await waitFor(() =>
-      expect(result.current.currentFields).toBe(fieldsToPersist),
-    );
-  });
-
-  it('should persist view fields', async () => {
-    const { result } = renderHook(
-      () => useViewBar({ viewBarId }),
-      renderHookConfig,
-    );
-
-    await act(async () => {
-      result.current.setCurrentViewId(mockedUuid);
-      result.current.setViewObjectMetadataId('newId');
-      await result.current.persistViewFields(fieldsToPersist);
-    });
-
-    const persistViewFieldsMutation = mocks[1];
-
-    await waitFor(() =>
-      expect(persistViewFieldsMutation.result).toHaveBeenCalled(),
-    );
-  });
-
-  describe('viewFilters', () => {
-    it('should load view filters', async () => {
-      const { result } = renderHook(() => {
-        const viewBar = useViewBar({ viewBarId });
-
-        const { currentViewFiltersState } = useViewScopedStates({
-          viewScopeId: viewBarId,
-        });
-        const currentViewFilters = useRecoilValue(currentViewFiltersState);
-
-        return {
-          viewBar,
-          currentViewFilters,
-        };
-      }, renderHookConfig);
-
-      expect(result.current.currentViewFilters).toStrictEqual([]);
-
-      await act(async () => {
-        result.current.viewBar.setAvailableFilterDefinitions([
-          filterDefinition,
-        ]);
-
-        await result.current.viewBar.loadViewFilters(
-          {
-            edges: [
-              {
-                node: viewFilter,
-                cursor: '',
-              },
-            ],
-            pageInfo: { hasNextPage: false, startCursor: '', endCursor: '' },
-          },
-          mockedUuid,
-        );
-        result.current.viewBar.setCurrentViewId(mockedUuid);
-      });
-
-      expect(result.current.currentViewFilters).toStrictEqual([viewFilter]);
-    });
-
-    it('should upsertViewFilter', async () => {
-      const { result } = renderHook(() => {
-        const viewBar = useViewBar({ viewBarId });
-
-        viewBar.setAvailableFilterDefinitions([filterDefinition]);
-
-        viewBar.loadViewFilters(
-          {
-            edges: [
-              {
-                node: viewFilter,
-                cursor: '',
-              },
-            ],
-            pageInfo: { hasNextPage: false, startCursor: '', endCursor: '' },
-          },
-          mockedUuid,
-        );
-        viewBar.setCurrentViewId(mockedUuid);
-
-        const { currentViewFiltersState } = useViewScopedStates({
-          viewScopeId: viewBarId,
-        });
-        const currentViewFilters = useRecoilValue(currentViewFiltersState);
-
-        return {
-          viewBar,
-          currentViewFilters,
-        };
-      }, renderHookConfig);
-
-      expect(result.current.currentViewFilters).toStrictEqual([viewFilter]);
-
-      const newFilters: Filter[] = [
-        {
-          fieldMetadataId: '113ea8f8-1908-4c9c-9984-3f23c96b92f5',
-          value: 'value',
-          displayValue: 'displayValue',
-          operand: ViewFilterOperand.IsNot,
-          definition: {
-            fieldMetadataId: 'id',
-            label: 'label',
-            iconName: 'icon',
-            type: 'TEXT',
-          },
-        },
-        {
-          fieldMetadataId: 'd9487757-183e-4fa0-a554-a980850cb23d',
-          value: 'value',
-          displayValue: 'displayValue',
-          operand: ViewFilterOperand.Contains,
-          definition: {
-            fieldMetadataId: 'id',
-            label: 'label',
-            iconName: 'icon',
-            type: 'TEXT',
-          },
-        },
-      ];
-
-      // upsert an existing filter
-      act(() => {
-        result.current.viewBar.upsertViewFilter(newFilters[0]);
-      });
-
-      expect(result.current.currentViewFilters).toStrictEqual([
-        { ...newFilters[0], id: viewFilter.id },
-      ]);
-
-      // upsert a new filter
-      act(() => {
-        result.current.viewBar.upsertViewFilter(newFilters[1]);
-      });
-
-      // expect currentViewFilters to contain both filters
-      expect(result.current.currentViewFilters).toStrictEqual([
-        { ...newFilters[0], id: viewFilter.id },
-        { ...newFilters[1], id: undefined },
-      ]);
-    });
-
-    it('should remove view filter', () => {
-      const { result } = renderHook(() => {
-        const viewBar = useViewBar({ viewBarId });
-
-        viewBar.setAvailableFilterDefinitions([filterDefinition]);
-
-        viewBar.loadViewFilters(
-          {
-            edges: [
-              {
-                node: viewFilter,
-                cursor: '',
-              },
-            ],
-            pageInfo: { hasNextPage: false, startCursor: '', endCursor: '' },
-          },
-          mockedUuid,
-        );
-        viewBar.setCurrentViewId(mockedUuid);
-
-        const { currentViewFiltersState } = useViewScopedStates({
-          viewScopeId: viewBarId,
-        });
-        const currentViewFilters = useRecoilValue(currentViewFiltersState);
-
-        return {
-          viewBar,
-          currentViewFilters,
-        };
-      }, renderHookConfig);
-
-      expect(result.current.currentViewFilters).toStrictEqual([viewFilter]);
-
-      // remove an existing filter
-      act(() => {
-        result.current.viewBar.removeViewFilter(
-          filterDefinition.fieldMetadataId,
-        );
-      });
-
-      expect(result.current.currentViewFilters).toStrictEqual([]);
-    });
-  });
-
-  describe('View Sorts', () => {
-    const currentViewId = 'ac8807fd-0065-436d-bdf6-94333d75af6e';
-
-    it('should load view sorts', async () => {
-      const { result } = renderHook(() => {
-        const viewBar = useViewBar({ viewBarId });
-
-        const { currentViewSortsState } = useViewScopedStates({
-          viewScopeId: viewBarId,
-        });
-        const currentViewSorts = useRecoilValue(currentViewSortsState);
-
-        return {
-          viewBar,
-          currentViewSorts,
-        };
-      }, renderHookConfig);
-
-      expect(result.current.currentViewSorts).toStrictEqual([]);
-
-      await act(async () => {
-        result.current.viewBar.setAvailableSortDefinitions([sortDefinition]);
-
-        await result.current.viewBar.loadViewSorts(
-          {
-            edges: [
-              {
-                node: viewSort,
-                cursor: '',
-              },
-            ],
-            pageInfo: { hasNextPage: false, startCursor: '', endCursor: '' },
-          },
-          currentViewId,
-        );
-        result.current.viewBar.setCurrentViewId(currentViewId);
-      });
-
-      expect(result.current.currentViewSorts).toStrictEqual([viewSort]);
-    });
-
-    it('should upsertViewSort', async () => {
-      const { result } = renderHook(() => {
-        const viewBar = useViewBar({ viewBarId });
-
-        viewBar.setAvailableSortDefinitions([sortDefinition]);
-
-        viewBar.loadViewSorts(
-          {
-            edges: [
-              {
-                node: viewSort,
-                cursor: '',
-              },
-            ],
-            pageInfo: { hasNextPage: false, startCursor: '', endCursor: '' },
-          },
-          mockedUuid,
-        );
-        viewBar.setCurrentViewId(mockedUuid);
-
-        const { currentViewSortsState } = useViewScopedStates({
-          viewScopeId: viewBarId,
-        });
-        const currentViewSorts = useRecoilValue(currentViewSortsState);
-
-        return {
-          viewBar,
-          currentViewSorts,
-        };
-      }, renderHookConfig);
-
-      expect(result.current.currentViewSorts).toStrictEqual([viewSort]);
-
-      const newSortFieldMetadataId = 'd9487757-183e-4fa0-a554-a980850cb23d';
-
-      const newSorts: Sort[] = [
-        {
-          fieldMetadataId: viewSort.fieldMetadataId,
-          direction: 'desc',
-          definition: sortDefinition,
-        },
-        {
-          fieldMetadataId: newSortFieldMetadataId,
-          direction: 'asc',
-          definition: {
-            ...sortDefinition,
-            fieldMetadataId: newSortFieldMetadataId,
-          },
-        },
-      ];
-
-      // upsert an existing sort
-      act(() => {
-        result.current.viewBar.upsertViewSort(newSorts[0]);
-      });
-
-      expect(result.current.currentViewSorts).toStrictEqual([
-        { ...newSorts[0], id: viewSort.id },
-      ]);
-
-      // upsert a new sort
-      act(() => {
-        result.current.viewBar.upsertViewSort(newSorts[1]);
-      });
-
-      // expect currentViewSorts to contain both sorts
-      expect(result.current.currentViewSorts).toStrictEqual([
-        { ...newSorts[0], id: viewSort.id },
-        { ...newSorts[1], id: undefined },
-      ]);
-    });
-
-    it('should remove view sort', () => {
-      const { result } = renderHook(() => {
-        const viewBar = useViewBar({ viewBarId });
-
-        viewBar.setAvailableSortDefinitions([sortDefinition]);
-
-        viewBar.loadViewSorts(
-          {
-            edges: [
-              {
-                node: viewSort,
-                cursor: '',
-              },
-            ],
-            pageInfo: { hasNextPage: false, startCursor: '', endCursor: '' },
-          },
-          mockedUuid,
-        );
-        viewBar.setCurrentViewId(mockedUuid);
-
-        const { currentViewSortsState } = useViewScopedStates({
-          viewScopeId: viewBarId,
-        });
-        const currentViewSorts = useRecoilValue(currentViewSortsState);
-
-        return {
-          viewBar,
-          currentViewSorts,
-        };
-      }, renderHookConfig);
-
-      expect(result.current.currentViewSorts).toStrictEqual([viewSort]);
-
-      // remove an existing sort
-      act(() => {
-        result.current.viewBar.removeViewSort(sortDefinition.fieldMetadataId);
-      });
-
-      expect(result.current.currentViewSorts).toStrictEqual([]);
-    });
-  });
-
-  it('should load view fields', async () => {
-    const fieldDefinition: ColumnDefinition<FieldMetadata> = {
-      size: 100,
-      position: 1,
-      fieldMetadataId: '12ecdf87-506f-44a7-98c6-393e5f05b225',
-      label: 'label',
-      iconName: 'icon',
-      type: 'TEXT',
-      metadata: {
-        placeHolder: 'placeHolder',
-        fieldName: 'fieldName',
-      },
-    };
-    const viewField: ViewField = {
-      id: '88930a16-685f-493b-a96b-91ca55666bba',
-      fieldMetadataId: '12ecdf87-506f-44a7-98c6-393e5f05b225',
-      position: 1,
-      isVisible: true,
-      size: 100,
-      definition: fieldDefinition,
-    };
-    const currentViewId = 'ac8807fd-0065-436d-bdf6-94333d75af6e';
-
-    const { result } = renderHook(() => {
-      const viewBar = useViewBar({ viewBarId });
-
-      const { currentViewFieldsState } = useViewScopedStates({
-        viewScopeId: viewBarId,
-      });
-      const currentViewFields = useRecoilValue(currentViewFieldsState);
-
-      return {
-        viewBar,
-        currentViewFields,
-      };
-    }, renderHookConfig);
-
-    expect(result.current.currentViewFields).toStrictEqual([]);
-
-    await act(async () => {
-      result.current.viewBar.setAvailableFieldDefinitions([fieldDefinition]);
-
-      await result.current.viewBar.loadViewFields(
-        {
-          edges: [
-            {
-              node: viewField,
-              cursor: '',
-            },
-          ],
-          pageInfo: { hasNextPage: false, startCursor: '', endCursor: '' },
-        },
-        currentViewId,
-      );
-      result.current.viewBar.setCurrentViewId(currentViewId);
-    });
-
-    expect(result.current.currentViewFields).toStrictEqual([viewField]);
   });
 
   it('should loadView', async () => {
